Update GameGateway
Move methods and variables to GameService
Connect the game status with the game pairing context
