+++ transcendence-app/src/game/infrastructure/db/memoryModels/index.ts
@@ -1,3 +1,2 @@
-export * from './games-ongoing.model';
 export * from './challenges-pending.model';
 export * from './types';
