+++ webapp/src/pages/GameQueuePage/GameQueuePage.tsx
@@ -18,19 +18,19 @@ import { PLAY_URL } from '../../shared/urls';
 import { GamePairingStatusDtoGameQueueStatusEnum } from '../../shared/generated';
 
 export default function GameQueuePage() {
-  const { gameQueueStatus, setGameCtx } = useGamePairing();
+  const { gameQueueStatus } = useGamePairing();
   const { goBack, navigate } = useNavigation();
   const quitAndGoBack = () => {
     if (gameQueueStatus === GamePairingStatusDtoGameQueueStatusEnum.Waiting) {
       socket.emit(gameQueueClientToServerWsEvents.gameQuitWaiting);
-      setGameCtx &&
-        setGameCtx({
-          gameQueueStatus: GamePairingStatusDtoGameQueueStatusEnum.None,
-          gameRoomId: null,
-        });
     }
     goBack();
   };
+
+  if (gameQueueStatus !== GamePairingStatusDtoGameQueueStatusEnum.Waiting) {
+    goBack();
+  }
+
   return (
     <div className="game-queue-page">
       <Header icon={IconVariant.ARROW_BACK} onClick={quitAndGoBack}>
@@ -52,11 +52,6 @@ export default function GameQueuePage() {
             gameQueueStatus === GamePairingStatusDtoGameQueueStatusEnum.Waiting
           ) {
             socket.emit(gameQueueClientToServerWsEvents.gameQuitWaiting);
-            setGameCtx &&
-              setGameCtx({
-                gameQueueStatus: GamePairingStatusDtoGameQueueStatusEnum.None,
-                gameRoomId: null,
-              });
           }
           navigate(PLAY_URL, { replace: true });
         }}
